backend/requirements.txt
fastapi
uvicorn
python-dotenv
redis
pydantic
yfinance
newsapi-python
requests
httpx
numpy
pandas
scikit-learn
xgboost
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pysqlite3-binary
huggingface_hub